注册逻辑优化--jackson

+++ src/main/java/com/dto/CookieDto.java
@@ -10,16 +10,16 @@ import java.util.Date;
  */
 public class CookieDto {
 
-    private long user_id; //用户id
+    private int user_id; //用户id
     private String cookie;  //cookie 由uuid产生，为唯一
     private String created_at; //创建时间
     private String updated_at; //修改时间
 
-    public long getUser_id() {
+    public int getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(long user_id) {
+    public void setUser_id(int user_id) {
         this.user_id = user_id;
     }
